requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.27.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
google-generativeai==0.3.1
openai==0.28.1
python-dotenv==1.0.0
pydantic==2.4.2
Pillow==10.0.1
